apps/requirements.txt
boto3==1.28.70
Pillow==10.1.0
streamlit==1.28.1